feat: Support reading Noir programs from in-mem binaries
In order to support WASM. And specially browser witgen, one of the first
steps is to actually eliminate the need to use `std::fs`. As aside from
`emscripten`-wasm target, the rest don't even have an implementation so
they will just fail in runtime.

This allows to load `Program` from binary therefore allowig in-browser
support for the crate.

This work is done towards https://github.com/privacy-scaling-explorations/sonobe/issues/147